+++ tests/test_eval_utils.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 import pytest
 
 from secheaders.exceptions import SecurityHeadersException
-from secheaders import utils
+from secheaders import eval_utils
 from secheaders.constants import EVAL_OK, EVAL_WARN
 
 from tests.constants import EXAMPLE_HEADERS
@@ -24,7 +24,7 @@ class TestUtils(TestCase):
             "base-uri": ["'self'"],
             "form-action": ["'self'"],
         }
-        res = utils.csp_parser(example_csp)
+        res = eval_utils.csp_parser(example_csp)
         assert res == expected_value
 
     def test_eval_csp(self) -> None:
@@ -32,7 +32,7 @@ class TestUtils(TestCase):
             "default-src 'none' *.example.com; script-src 'self' src.example.com 'unsafe-inline'; connect-src 'self';"
             "img-src *; style-src 'self'; base-uri 'self';form-action 'self'"
         )
-        res = utils.eval_csp(unsafe_csp)
+        res = eval_utils.eval_csp(unsafe_csp)
         expected_value = (
             EVAL_WARN,
             ["Unsafe source 'unsafe-inline' in directive script-src"]
@@ -41,15 +41,15 @@ class TestUtils(TestCase):
 
         safe_csp = "default-src 'self'; img-src 'self' cdn.example.com;"
         expected_value = (EVAL_OK, [])
-        res = utils.eval_csp(safe_csp)
+        res = eval_utils.eval_csp(safe_csp)
         assert res == expected_value
 
     def test_eval_version_info(self) -> None:
         nginx_banner_warn = 'nginx 1.17.10 (Ubuntu)'
         nginx_banner_ok = 'nginx'
-        res = utils.eval_version_info(nginx_banner_warn)
+        res = eval_utils.eval_version_info(nginx_banner_warn)
         assert res == (EVAL_WARN, [])
-        res = utils.eval_version_info(nginx_banner_ok)
+        res = eval_utils.eval_version_info(nginx_banner_ok)
         assert res == (EVAL_OK, [])
 
     def test_permissions_policy_parser(self) -> None:
@@ -61,7 +61,7 @@ class TestUtils(TestCase):
             'picture-in-picture': [],
             'camera': ['*'],
         }
-        res = utils.permissions_policy_parser(example_pp)
+        res = eval_utils.permissions_policy_parser(example_pp)
         assert expected_value == res
 
     def test_eval_permissions_policy(self) -> None:
@@ -71,11 +71,11 @@ class TestUtils(TestCase):
                 "Privacy-sensitive feature 'microphone' not defined in permission-policy, always allowed.",
                 "Privacy-sensitive feature 'payment' not defined in permission-policy, always allowed.",
         ])
-        res = utils.eval_permissions_policy(unsafe_pp)
+        res = eval_utils.eval_permissions_policy(unsafe_pp)
         assert res == expected_value
         safe_pp = "geolocation=(src), camera=(), microphone=(), payment=()"
         expected_value = EVAL_OK, []
-        res = utils.eval_permissions_policy(safe_pp)
+        res = eval_utils.eval_permissions_policy(safe_pp)
         assert res == expected_value
 
     def test_eval_headers(self) -> None:
@@ -84,8 +84,8 @@ class TestUtils(TestCase):
             'x-xss-protection': '1;',
         }
         with pytest.raises(SecurityHeadersException, match=r"Headers not fetched successfully"):
-            utils.analyze_headers({})
+            eval_utils.analyze_headers({})
 
-        res = utils.analyze_headers(fetched_headers)
+        res = eval_utils.analyze_headers(fetched_headers)
 
         assert res == EXAMPLE_HEADERS
